use torchaudio argument instead of separate top-level NeuralModule
