+++ nemo/collections/asr/modules/__init__.py
@@ -14,7 +14,6 @@
 
 from nemo.collections.asr.modules.audio_preprocessing import (
     AudioToMelSpectrogramPreprocessor,
-    AudioToMelSpectrogramPreprocessorX,
     AudioToMFCCPreprocessor,
     CropOrPadSpectrogramAugmentation,
     MaskedPatchAugmentation,
